requirements.txt
# Core quantum computing and machine learning
pennylane>=0.28.0
pennylane-lightning>=0.28.0
torch>=1.10.0
numpy<2
scipy>=1.10,<1.14

# Chemistry and molecular processing
rdkit-pypi==2023.9.5
selfies>=2.0.0

# Data processing and analysis
pandas>=1.2.0
matplotlib>=3.3.0
seaborn>=0.11.0
tqdm>=4.60.0

# Optimization and hyperparameter tuning
optuna>=2.10.0

# Molecular property calculation and drug discovery
syba>=0.0.5

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
isort>=5.10.0
flake8>=5.0.0

# Documentation
jupyter>=1.0.0
jupyter-book>=0.13.0
sphinx>=4.0.0

# Configuration management
pyyaml>=6.0
hydra-core>=1.2.0

# Utilities
click>=8.0.0
python-dotenv>=0.19.0

# Optional: GPU support for PyTorch (uncomment if needed)
# torch-audio>=0.12.0
# torch-vision>=0.13.0

# Optional: Additional quantum backends
# pennylane-qiskit>=0.30.0
# pennylane-cirq>=0.30.0

# Optional: Distributed computing
# dask>=2022.0.0
# ray>=2.0.0
